PYIC-8284 Extract more helpers
