+++ journey-map/src/helpers/deep-clone.ts
@@ -1,0 +1,2 @@
+// Simple deep clone - N.B. this will only work with pure JSON objects
+export const deepCloneJson = <T>(obj: T): T => JSON.parse(JSON.stringify(obj));
