+++ journey-map/src/render.ts
@@ -1,10 +1,11 @@
-import { getNestedJourneyStates, resolveEventTargets } from "./helpers.js";
-import {
-  JourneyEvent,
-  JourneyMap,
-  JourneyState,
-  NestedJourneyMap,
-} from "./types.js";
+import { getNestedJourneyStates } from "./helpers/uplift-nested.js";
+import { resolveVisibleEventTargets } from "./helpers/event-resolver.js";
+import { expandNestedJourneys } from "./helpers/expand-nested.js";
+import { expandParents } from "./helpers/expand-parents.js";
+import { RenderOptions } from "./helpers/options.js";
+import { findOrphanStates } from "./helpers/orphans.js";
+import { JourneyMap, JourneyState, NestedJourneyMap } from "./types.js";
+import { deepCloneJson } from "./helpers/deep-clone.js";
 
 const topDownJourneys = ["INITIAL_JOURNEY_SELECTION"];
 const errorJourneys = ["TECHNICAL_ERROR"];
@@ -13,143 +14,16 @@ const failureJourneys = ["INELIGIBLE", "FAILED"];
 const JOURNEY_CONTEXT_TRANSITION_CLASSNAME = "journeyCtxTransition";
 const MITIGATIONS_TRANSITION_CLASSNAME = "mitigationTransition";
 
-// Simple deep clone - N.B. this will only work with pure JSON objects
-const deepCloneJson = <T>(obj: T): T => JSON.parse(JSON.stringify(obj));
-
-// Expand out parent states
-// Will also search 'otherStates' (e.g. for nested journeys)
-const expandParents = (
-  journeyStates: Record<string, JourneyState>,
-  otherStates: Record<string, JourneyState>,
-): void => {
-  const parentStates: string[] = [];
-  Object.entries(journeyStates).forEach(([state, definition]) => {
-    if (definition.parent) {
-      const parent =
-        journeyStates[definition.parent] ?? otherStates[definition.parent];
-      if (!parent) {
-        console.warn(`Missing parent ${definition.parent} of state ${state}`);
-      } else {
-        definition.events = {
-          ...parent.events,
-          ...definition.events,
-        };
-        journeyStates[state] = { ...parent, ...definition };
-        parentStates.push(definition.parent);
-      }
-    }
-  });
-  // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-  parentStates.forEach((state) => delete journeyStates[state]);
-};
-
-const mapTargetStateToExpandedState = (
-  eventDef: JourneyEvent,
-  subJourneyState: string,
-  formData: FormData,
-): void => {
-  // Map target states to expanded states
-  resolveEventTargets(eventDef, formData).forEach((targetDef) => {
-    if (targetDef.targetState && !targetDef.targetJourney) {
-      targetDef.targetState = `${subJourneyState}/${targetDef.targetState}`;
-    }
-  });
-};
-
-// Expand out nested states
-const expandNestedJourneys = (
-  journeyMap: Record<string, JourneyState>,
-  subjourneys: Record<string, NestedJourneyMap>,
-  formData: FormData,
-): void => {
-  Object.entries(journeyMap).forEach(([state, definition]) => {
-    if (definition.nestedJourney && subjourneys[definition.nestedJourney]) {
-      const subJourneyState = state;
-      // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-      delete journeyMap[subJourneyState];
-      const subjourney = subjourneys[definition.nestedJourney];
-
-      // Expand out each of the nested states
-      Object.entries(subjourney.nestedJourneyStates).forEach(
-        ([nestedState, nestedDefinition]) => {
-          // Copy to avoid mutating different versions of the expanded definition
-          const expandedDefinition = deepCloneJson(nestedDefinition);
-
-          Object.entries(expandedDefinition.events || {}).forEach(
-            ([evt, eventDef]) => {
-              mapTargetStateToExpandedState(
-                eventDef,
-                subJourneyState,
-                formData,
-              );
-
-              // Map exit events to targets in the parent definition
-              const exitEvent = eventDef.exitEventToEmit;
-              if (exitEvent) {
-                delete eventDef.exitEventToEmit;
-                if (definition.exitEvents?.[exitEvent]) {
-                  Object.assign(eventDef, definition.exitEvents[exitEvent]);
-                } else {
-                  console.warn(
-                    `Unhandled exit event from ${subJourneyState}:`,
-                    exitEvent,
-                  );
-                  delete expandedDefinition.events?.[evt];
-                }
-              }
-            },
-          );
-
-          journeyMap[`${subJourneyState}/${nestedState}`] = expandedDefinition;
-        },
-      );
-
-      // Make a copy of the entry events to avoid mutating the original
-      const entryEvents = deepCloneJson(subjourney.entryEvents);
-      // Update entry events on other states to expanded states
-      Object.entries(entryEvents).forEach(([entryEvent, entryEventDef]) => {
-        mapTargetStateToExpandedState(entryEventDef, subJourneyState, formData);
-
-        Object.values(journeyMap).forEach((journeyDef) => {
-          if (journeyDef.events?.[entryEvent]) {
-            resolveEventTargets(journeyDef.events[entryEvent], formData)
-              .filter(
-                (t) => t.targetState === subJourneyState && !t.targetEntryEvent,
-              )
-              .forEach((t) => {
-                Object.assign(t, entryEventDef);
-              });
-          }
-
-          // Resolve targets with a `targetEntryEvent` override
-          Object.values(journeyDef.events ?? {}).forEach((eventDef) => {
-            resolveEventTargets(eventDef, formData)
-              .filter(
-                (t) =>
-                  t.targetState === subJourneyState &&
-                  t.targetEntryEvent === entryEvent,
-              )
-              .forEach((t) => {
-                Object.assign(t, entryEventDef);
-                delete t.targetEntryEvent;
-              });
-          });
-        });
-      });
-    }
-  });
-};
-
 interface TransitionsOutput {
   transitionsMermaid: string;
   states: string[];
 }
 
 // Render the transitions into mermaid, while tracking the states traced from the initial states
-// This allows us to skip
+// This allows us to skip unreachable states
 const renderTransitions = (
   journeyStates: Record<string, JourneyState>,
-  formData: FormData,
+  options: RenderOptions,
 ): TransitionsOutput => {
   // Initial states have no response or nested journey
   const initialStates = Object.keys(journeyStates).filter(
@@ -165,7 +39,7 @@ const renderTransitions = (
 
     const eventsByTarget: Record<string, string[]> = {};
     Object.entries(events).forEach(([eventName, def]) => {
-      const resolvedEventTargets = resolveEventTargets(def, formData);
+      const resolvedEventTargets = resolveVisibleEventTargets(def, options);
 
       for (const resolvedTarget of resolvedEventTargets) {
         const {
@@ -185,13 +59,13 @@ const renderTransitions = (
 
         if (
           errorJourneys.includes(targetJourney as string) &&
-          !formData.getAll("otherOption").includes("includeErrors")
+          !options.includeErrors
         ) {
           continue;
         }
         if (
           failureJourneys.includes(targetJourney as string) &&
-          !formData.getAll("otherOption").includes("includeFailures")
+          !options.includeFailures
         ) {
           continue;
         }
@@ -351,39 +225,6 @@ const renderStates = (
   return mermaids.join("\n");
 };
 
-const resolveAllPossibleEventTargets = (
-  eventDefinition: JourneyEvent,
-): string[] => [
-  eventDefinition.targetState,
-  ...Object.values(eventDefinition.checkIfDisabled || {}).flatMap(
-    resolveAllPossibleEventTargets,
-  ),
-  ...Object.values(eventDefinition.checkFeatureFlag || {}).flatMap(
-    resolveAllPossibleEventTargets,
-  ),
-];
-
-const calcOrphanStates = (
-  journeyMap: Record<string, JourneyState>,
-): string[] => {
-  const targetedStates = [
-    ...Object.values(journeyMap).flatMap((stateDefinition) => [
-      ...Object.values(stateDefinition.events || {}).flatMap(
-        resolveAllPossibleEventTargets,
-      ),
-      ...Object.values(stateDefinition.exitEvents || {}).flatMap(
-        resolveAllPossibleEventTargets,
-      ),
-    ]),
-  ];
-
-  const uniqueTargetedStates = [...new Set(targetedStates)];
-
-  return Object.keys(journeyMap).filter(
-    (state) => !uniqueTargetedStates.includes(state),
-  );
-};
-
 const getMermaidGraph = (
   graphDirection: "TD" | "LR",
   statesMermaid: string,
@@ -406,7 +247,7 @@ export const render = (
   selectedJourney: string,
   journeyMap: JourneyMap,
   nestedJourneys: Record<string, NestedJourneyMap>,
-  formData = new FormData(),
+  options: RenderOptions,
 ): string => {
   const isNestedJourney = selectedJourney in nestedJourneys;
   const direction = topDownJourneys.includes(selectedJourney) ? "TD" : "LR";
@@ -418,28 +259,23 @@ export const render = (
       : journeyMap.states,
   );
 
-  if (
-    !isNestedJourney &&
-    formData.getAll("otherOption").includes("expandNestedJourneys")
-  ) {
+  if (!isNestedJourney && options.expandNestedJourneys) {
     // Expand nested journeys first, to allow for two levels of nesting
     Object.values(nestedJourneys).forEach((nestedJourney) =>
       expandNestedJourneys(
         nestedJourney.nestedJourneyStates,
         nestedJourneys,
-        formData,
+        options,
       ),
     );
-    expandNestedJourneys(journeyStates, nestedJourneys, formData);
+    expandNestedJourneys(journeyStates, nestedJourneys, options);
   }
 
   expandParents(journeyStates, journeyMap.states);
 
-  const { transitionsMermaid, states } = formData
-    .getAll("otherOption")
-    .includes("onlyOrphanStates")
-    ? { transitionsMermaid: "", states: calcOrphanStates(journeyStates) }
-    : renderTransitions(journeyStates, formData);
+  const { transitionsMermaid, states } = options.onlyOrphanStates
+    ? { transitionsMermaid: "", states: findOrphanStates(journeyStates) }
+    : renderTransitions(journeyStates, options);
 
   const statesMermaid = renderStates(journeyStates, states);
